+++ src/components/relative/relative-date-range.component.jsx
@@ -142,6 +142,7 @@ export default class RelativeDateRange extends React.PureComponent {
               clearable={false}
               onChange={this.handleStartDateChange}
               options={startDateOptions}
+              placeholder={translations.startDatePlaceholder}
               value={startDate}
             />
           </Content.InputColumn>
@@ -158,6 +159,7 @@ export default class RelativeDateRange extends React.PureComponent {
               clearable={false}
               onChange={this.handleEndDateChange}
               options={endDateOptions}
+              placeholder={translations.endDatePlaceholder}
               value={endDate}
             />
           </Content.InputColumn>